refactor: refactored the structure of the codebase based on the instructor feedback
