+++ src/main/java/domain/CardServices.java
@@ -1,4 +1,4 @@
-package ui;
+package domain;
 
 import java.util.List;
 import java.util.Random;
