+++ src/main/java/domain/model/GameSetupModel.java
@@ -1,4 +1,4 @@
-package ui.model;
+package domain.model;
 
 import java.util.ArrayList;
 import java.util.List;
