+++ src/test/java/ui/CardServicesTest.java
@@ -3,6 +3,7 @@ package ui;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
+import domain.CardServices;
 import java.util.Random;
 import org.easymock.EasyMock;
 import org.junit.jupiter.api.Test;
